modify-records

+++ server/records.cpp
@@ -69,16 +69,16 @@ QVector<QString> Records::getSimpleBills(int roomId)
     QString str;
     //QDateTime max = QDateTime::fromString("00-00-00 00:00:00");
     //int maxi = 0;
-    //uint stime;
-    //uint etime;
-    //uint htime;
-    int tRet;
+    //int64_t stime;
+    //int64_t etime;
+    //int64_t htime;
+    int64_t tRet;
     if(!billing.empty())
     {
-        uint stimm;
-        uint etimm;
-        uint htimm;
-        int n;
+        int64_t stimm;
+        int64_t etimm;
+        int64_t htimm;
+        int64_t n;
         double cost=0;
         stimm = roomss.at(roomId).start.toTime_t();
         htimm = roomss.at(roomId).duration.toTime_t();
@@ -120,25 +120,25 @@ QVector<QString> Records::getReportForm(QDateTime start , QDateTime end)
   // TODO
     QVector<QString> data;
     QString str;
-    uint stim;
-    uint etim;
-    uint etim2;
-    int tRet,tRet2;
+    int64_t stim;
+    int64_t etim;
+    int64_t etim2;
+    int64_t tRet,tRet2;
     int *op;//开关次数
     int *temp;//调温
     int *speed;//调风
     int *record;//详单
-    int *duratio;//时长
-    int *fee;//总消费
+    int64_t *duratio;//时长
+    double *fee;//总消费
     op=(int*)calloc(rooms.size(),sizeof(int));
     temp=(int*)calloc(rooms.size(),sizeof(int));
     speed=(int*)calloc(rooms.size(),sizeof(int));
     record=(int*)calloc(rooms.size(),sizeof(int));
-    duratio=(int*)calloc(rooms.size(),sizeof(int));
-    fee=(int*)calloc(rooms.size(),sizeof(int));
+    duratio=(int64_t*)calloc(rooms.size(),sizeof(int));
+    fee=(double*)calloc(rooms.size(),sizeof(int));
     for(int j=0; j<billingss.length(); ++j)
     {
-        int dus;//持续时间
+        int64_t dus;//持续时间
         stim = start.toTime_t();
         etim = billingss.at(j).start.toTime_t();
         etim2 = end.toTime_t();
@@ -183,9 +183,9 @@ QVector<QString> Records::getReportForm(QDateTime start , QDateTime end)
             str.append(QString::number(i));
             str.append(" 使用时长：");
             //空调运行时间
-            int du;
+            int64_t du;
             du = duratio[i];
-            int h = du / 3600;
+            int64_t h = du / 3600;
             int h1 = du % 3600;
             str.append(QString::number(h));
             str.append("时");
@@ -227,7 +227,7 @@ QString Records::calcDurationStr(long long duration)
     else
     {
         long long d = duration;
-        int h, m;
+        int64_t h, m;
         h = d / 3600;
         d = d % 3600;
         m = d / 60;
